Changed third-person camera + added wall jump / wall sliding
Removed first person view (it was too boring + couldn't see animations)
Changed the camera to a Cinemachine camera which allows more freedom.
Added parts of the wall jump script which allows the character to slide down a wall at a reduced speed and jump away from the wall while in mid air + added animations for it.